+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added Dwarfing enchantment
   * Increased mining speed the deeper you mine
+* Adrenaline max bonus is now at 35% durability left instead of 0%
 * Mining Speed tooltip no longer takes into account player's mining speed changes (such as effects or being in water / off ground)
 
 ## 1.31.0